# DLTE-FrontEnd/DLTE-FrontEnd-Tasks/DLTE-FrontEnd-Task-1292/files/api-docs final.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/update/limit":{"put":{"tags":["update-card-limit"],"summary":"Updating the Debit Card Limit","operationId":"updateLimit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DebitCard"}}},"required":true},"responses":{"400":{"description":"Debit card limit update failed","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"404":{"description":"Account is not Active","content":{"*/*":{"schema":{"type":"string"}}}},"422":{"description":" Please provide the correct debit card details","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Debit card limit updated successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/profile/register":{"post":{"tags":["card-security-api"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardSecurity"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSecurity"}}}}}}}},"components":{"schemas":{"DebitCard":{"required":["accountNumber","customerId","debitCardCvv","debitCardExpiry","debitCardNumber","debitCardPin","debitCardStatus","domesticLimit","internationalLimit"],"type":"object","properties":{"debitCardNumber":{"type":"integer","format":"int64"},"accountNumber":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int32"},"debitCardCvv":{"type":"integer","format":"int32"},"debitCardPin":{"type":"integer","format":"int32"},"debitCardExpiry":{"type":"string","format":"date-time"},"debitCardStatus":{"pattern":"^(active|inactive)$","type":"string"},"domesticLimit":{"type":"number","format":"double"},"internationalLimit":{"type":"number","format":"double"}}},"CardSecurity":{"type":"object","properties":{"customerName":{"type":"string"},"customerAddress":{"type":"string"},"customerStatus":{"type":"string"},"customerContact":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"maxAttempt":{"type":"integer","format":"int32"},"attempts":{"type":"integer","format":"int32"},"enabled":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}}}}}